product-service/requirements.txt
flask
psycopg2-binary
redis

